Supprime les topics et posts de Typesense lors du déplacement vers un forum privé

+++ zds/search/utils.py
@@ -287,6 +287,8 @@ class SearchIndexManager:
 
         doc_type = document.get_search_document_type()
         doc_id = document.search_engine_id
+        if doc_id is None:
+            return
         answer = self.engine.collections[doc_type].documents[doc_id].delete()
         if "id" not in answer or answer["id"] != doc_id:
             self.logger.warn(f"Error when deleting: {answer}.")
